Addresses and prefixes now represented as bitmasked uint32/64
Huge performance gains by working entirely with bitmasking for
addresses and node prefixes. IPv4 addresses are represented as
uint32, and IPv6 are now represented as a pair of uint64's.

We no longer have a generic tree that handles both IPv4 and
IPv6. There's now a 32-bit IPv4 tree, and a 128-bit IPv6 tree.
This was done for performance - the IPv4 tree will perform a
little better, since its underlying representation and
operations are simpler.

The trees now support storing all of the built-in types, as
well as slices of these types. This is done through code
generation. Having specifically-typed trees proved to be
much more performant than just supporting interface{} trees.
The code in the `template` directory is the source. The other
directories are generated with `make`.