Generate code using Go generics (#20)
* Generate code using Go generics

The code mutation is a bit less pretty but it still works well. Tests
are kept in this case as they are easy to mutate to check that
everything is working.

However, this requires bumping the minimal Go version to 1.18. For
clarity, generated code will be in a separate commit.

* Add generated code with generics

* Bump golangci/golangci-lint-action version for Go 1.18 compatibility