Add `Delete*()` methods to iterator (#26)
This is useful to remove tags matching a certain criteria in the whole
tree. For example, when implementing a RIB, we may want to remove tags
associated with a down peer.

The tree compaction is extracted into its own function and returns an
hint on how compaction was done. This hint is then used by the
`Delete*()` functions to update the current iterator.