Add TagsFromRoot
Add a method to the tree iterator that
returns all tags from all the node from the root of the tree
to the current leaf, in that order.

This is usefull to find the most relevant tag, when it's not defined
on the node itself. Similar to FindDeepestTags, but for iter.